requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bitsandbytes==0.43.3
boto3==1.35.1
botocore==1.35.1
cffi==1.17.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
encodec==0.1.1
exceptiongroup==1.2.2
fastapi==0.112.1
ffmpy==0.4.0
fonttools==4.53.1
fsspec==2024.6.1
funcy==2.0
gradio==4.33.0
gradio_client==0.17.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
importlib_metadata==8.4.0
importlib_resources==6.4.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langid==1.1.6
llvmlite==0.41.1
markdown-it-py==3.0.0
matplotlib==3.7.5
mdurl==0.1.2
mkl-service==2.4.0
more-itertools==10.4.0
narwhals==1.5.0
nltk==3.9.1
numba==0.58.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
OpenCC==1.1.9
orjson==3.10.7
packaging==24.1
pandas==2.0.3
pkgutil_resolve_name==1.3.10
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
referencing==0.35.1
regex==2024.7.24
rich==13.7.1
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.6.1
s3transfer==0.10.2
safetensors==0.4.4
scipy==1.10.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.38.2
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers @ git+https://github.com/huggingface/transformers.git@c63a3d0f1791e018de447ac570fc7029d1ea19bd
triton==2.3.1
typer==0.12.4
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
websockets==11.0.3
zipp==3.20.0
